add endpoints to controller
